# backend/src/main/java/com/visitor/service/auth/dto/RegisterRequest.java
package com.visitor.service.auth.dto;

import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;

public record RegisterRequest(
        @NotBlank @Size(min = 3, max = 64) String username,
        @NotBlank @Size(min = 6, max = 64) String password,
        @Size(max = 128) String displayName
) {
}